# Optimize familyCrible creation