--- a/go.mod
+++ b/go.mod
@@ -1,8 +1,9 @@
 module github.com/go-text/typesetting
 
-go 1.17
+go 1.18
 
 require (
 	github.com/go-text/typesetting-utils v0.0.0-20230616150549-2a7df14b6a22
+	golang.org/x/exp v0.0.0-20230817173708-d852ddb80c63
 	golang.org/x/image v0.3.0
 )