Generally speaking `container/list` is usually not the best choice for constructing things. This replaces that list/container with a regular slice and gets a significant performance win.

There probably is a block linked list or rope like structure that would probably work even better.