@egonelbre 2023-09-01
@whereswaldon ok, I created non-generic funcs of them