@benoitkugler 2023-09-01
Could you provide some benchmark or evidence that this is actually a performance gain? Thank you!