@egonelbre 2023-09-02
The second commit contained the benchmark:

```
  goos: windows
  goarch: amd64
  pkg: github.com/go-text/typesetting/fontscan
  cpu: AMD Ryzen Threadripper 2950X 16-Core Processor
                                │  base~.log  │             slice~.log              │
                                │   sec/op    │   sec/op     vs base                │
  ResolveFont-32                  21.57µ ± 2%   21.91µ ± 4%   +1.55% (p=0.035 n=10)
  SetQuery-32                     53.06n ± 2%   52.75n ± 3%        ~ (p=1.000 n=10)
  NewFamilyCrible-32              40.83µ ± 1%   21.96µ ± 1%  -46.22% (p=0.000 n=10)
  ScriptSet/naive-32              112.1m ± 1%   106.0m ± 1%   -5.46% (p=0.000 n=10)
  ScriptSet/from_rune_pages-32    4.140m ± 6%   3.894m ± 1%   -5.94% (p=0.000 n=10)
  ScriptSet/from_rune_ranges-32   1.767m ± 1%   1.727m ± 1%   -2.27% (p=0.000 n=10)
  ScanFonts-32                    430.1m ± 3%   427.4m ± 2%        ~ (p=0.579 n=10)
  geomean                         556.3µ        499.2µ       -10.26%
```

But, did you mean something additional to it?