@andydotxyz 2023-09-03
> > I'm not quite sure, but what's the minimum version this package should support?
> 
> We are limited to supporting the oldest version of Go that Fyne supports. That was Go 1.14, but _may_ be Go 1.17 now. @andydotxyz, did that transition already happen? Either way, we unfortunately cannot use generics.

Yes, we released the update to require Go 1.17 just 2 days ago :)