@egonelbre 2023-09-04
> This looks good, but should there not be tests for the 2 new methods that replace List methods?

I can add them, but I was partly hoping that they will be replaced by `slices.*`, so there wouldn't be a need to maintain them. Is there a timeline for fyne upgrading to 1.18? But, if not then I can add some tests.