@egonelbre 2023-09-04
@andydotxyz, I added tests, and of course, found a bug :)