@andydotxyz 2023-09-04
> Is there a timeline for fyne upgrading to 1.18? But, if not then I can add some tests.

We are discussing moving to 1.19 in the next major release - no ETA yet but I would guess January '24.